requirements.txt
flask==3.1.0
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
python-dotenv==1.0.1
bcrypt==4.2.1
